Improved performance:

+++ Assignment_3/include/properties.h
@@ -3,7 +3,7 @@
 
 
 
-int compute_energy(std::vector<std::vector<int> >& lattice);
-void write_properties(std::vector<int> energies, std::string outfile);
-void write_energies(std::vector<int> energies, std::string outfile);
+double compute_energy(std::vector<std::vector<int> >& lattice);
+void write_properties(std::vector<double> energies, std::string outfile);
+void write_energies(std::vector<double> energies, std::string outfile);
 void write_configuration(std::vector<std::vector<int> >& lattice, std::string outfile);
